Services and Repositories added for admin

+++ src/main/java/com/exam/system/dtos/ModuleRequestDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ModuleRequestDto {
-    private int examId;
+    private long examId;
     private String name;
     private String description;
 }
